feat: AuthProvider를 삭제 하고 react query로 인증 데이터 관리
